Update Final Parsing Logic (#14)
Resolves #13

### Summary
This PR edits the parsing logic so that a section has a final if it satisfies ALL of the following conditions:
1. Has a "Lecture" designation for its schedule type
2. Has at least 2 credit hours
3. Not a VIP course

In addition, this PR also updates how credit hours are parsed from Oscar. Apparently, Oscar's data entry is not very consistent. The current field that the crawler uses to obtain credit hours is unreliable and can be null/undefined. As a result, we would need to combine scraping data from another field to eliminate all null credit hours as much as possible. If null credit hours still exist after this, it will default to 0.

---------

Co-authored-by: Hailey Ho <haileyho@BMESNEL-MBP14-A-2.local>